Extract Seo component, mock follow link
