+++ src/components/seo/index.ts
@@ -1,0 +1,3 @@
+import Seo from './seo';
+
+export default Seo;
